Update README.md
It may be different for others but for me, I am approaching the use of the plugin as a first time user of drools so I would appreciate the links to the drools documentation.  May be worthwhile emphasizing at the start of the document that first time users of Drools might be served best by reading chapters 6 and 7 before using the plugin?